+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2022-07-07
+### Added
+- Add `MagicBell.connect` and `MagicBell.disconnect` methods for more ergonomic connection management.
+
+### Fixed
+- Reorder `Notification` types to avoid forward ref issue in pydantic validation
+
 ## [1.0.0] - 2022-07-05
 First open source release of the SDK.